requirements.txt
flask
flask-marshmallow
mysql-connector-python